test: move new state and event filtering tests into existing suites
- fold the late-subscriber source config listener tests into
  EventFilteringPluginTests and the observeDispatched/atomicity tests into
  StateManagementTests, instead of keeping them in separate files.
- reuse the existing MockStateAction rather than declaring bespoke actions.

Scanned-by: gitleaks 8.29.0
