+++ Tests/RudderStackAnalyticsTests/Plugins/DeviceMode/EventFilteringPluginTests.swift
@@ -348,10 +348,95 @@ struct EventFilteringPluginTests {
         ]
         
         let results = events.map { plugin.intercept(event: $0) }
-        
+
         #expect(results[0] == nil)
         #expect(results[1] != nil)
         #expect(results[2] == nil)
         #expect(results[3] != nil)
+    }
+
+    // MARK: - Source Config Listener Tests
+
+    /*
+     Unlike the tests above, these drive the plugin through `sourceConfigState` rather than the
+     `updateConfiguration` hook, so they cover the observation path itself. They guard the ordering
+     bug where an integration added after the source config had already been dispatched left its
+     filtering list empty and silently forwarded denied events.
+     */
+
+    /// "Bing Ads" in mock_source_config.json uses eventFilteringOption = "whitelistedEvents".
+    private var sourceConfigDestinationKey: String { "Bing Ads" }
+    private var whitelistedEvent: String { "form_submit" }
+    private var nonWhitelistedEvent: String { "Track Event 3" }
+
+    /// Returns an analytics instance whose own cached/network source config fetch has settled,
+    /// so that any dispatch afterwards is the one under test.
+    private func settledAnalytics() async -> Analytics {
+        let analytics = MockProvider.createMockAnalytics()
+        await runAfter(1.0) {}
+        return analytics
+    }
+
+    @Test("Given a source config dispatched before setup, When intercepting track events, Then filtering should be applied")
+    func testLateSubscriberAppliesFiltering() async {
+        let analytics = await settledAnalytics()
+        guard let sourceConfig = MockProvider.sourceConfiguration else {
+            #expect(Bool(false), "sourceConfiguration is nil")
+            return
+        }
+
+        // Source config is already in the state before the plugin subscribes, which is what happens
+        // when an integration is added after the source config has been fetched.
+        analytics.sourceConfigState.dispatch(action: UpdateSourceConfigAction(updatedSourceConfig: sourceConfig))
+
+        let plugin = EventFilteringPlugin(key: sourceConfigDestinationKey)
+        plugin.setup(analytics: analytics)
+
+        await runAfter(0.5) {
+            #expect(plugin.intercept(event: TrackEvent(event: whitelistedEvent)) != nil)
+            #expect(plugin.intercept(event: TrackEvent(event: nonWhitelistedEvent)) == nil)
+        }
+    }
+
+    @Test("Given a source config dispatched before setup and an identical refresh, When intercepting a track event, Then filtering should still be applied")
+    func testLateSubscriberWithIdenticalRefreshAppliesFiltering() async {
+        let analytics = await settledAnalytics()
+        guard let sourceConfig = MockProvider.sourceConfiguration else {
+            #expect(Bool(false), "sourceConfiguration is nil")
+            return
+        }
+
+        analytics.sourceConfigState.dispatch(action: UpdateSourceConfigAction(updatedSourceConfig: sourceConfig))
+
+        let plugin = EventFilteringPlugin(key: sourceConfigDestinationKey)
+        plugin.setup(analytics: analytics)
+
+        // A refresh returning an identical config is de-duplicated on updatedAt, so the plugin must
+        // already be configured from the value it received on subscription.
+        await runAfter(0.2) {}
+        analytics.sourceConfigState.dispatch(action: UpdateSourceConfigAction(updatedSourceConfig: sourceConfig))
+
+        await runAfter(0.5) {
+            #expect(plugin.intercept(event: TrackEvent(event: nonWhitelistedEvent)) == nil)
+        }
+    }
+
+    @Test("Given a source config dispatched after setup, When intercepting track events, Then filtering should be applied")
+    func testEarlySubscriberAppliesFiltering() async {
+        let analytics = await settledAnalytics()
+        guard let sourceConfig = MockProvider.sourceConfiguration else {
+            #expect(Bool(false), "sourceConfiguration is nil")
+            return
+        }
+
+        let plugin = EventFilteringPlugin(key: sourceConfigDestinationKey)
+        plugin.setup(analytics: analytics)
+
+        analytics.sourceConfigState.dispatch(action: UpdateSourceConfigAction(updatedSourceConfig: sourceConfig))
+
+        await runAfter(0.5) {
+            #expect(plugin.intercept(event: TrackEvent(event: whitelistedEvent)) != nil)
+            #expect(plugin.intercept(event: TrackEvent(event: nonWhitelistedEvent)) == nil)
+        }
     }
 }
